requirements.txt
requests~=2.32.3
PySide6~=6.7.3
pyinstaller
beautifulsoup4
selenium
pandas~=2.2.3
python-dotenv~=1.0.1
openpyxl
gspread~=6.1.4
oauth2client
yfinance~=0.2.48
finnhub-python
google-auth
google-auth-oauthlib
